+++ src/test/java/com/camping/admin/CucumberTestRunner.java
@@ -1,4 +1,4 @@
-package com.camping;
+package com.camping.admin;
 
 import org.junit.platform.suite.api.SelectClasspathResource;
 import org.junit.platform.suite.api.Suite;
